requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
bidict==0.23.1
ccxt==4.4.6
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
dateparser==1.2.0
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
multidict==6.1.0
numpy==2.1.1
pandas==2.2.3
pandas_ta==0.3.14b0
psycopg2-binary==2.9.9
pycares==4.4.0
pycparser==2.22
pycryptodome==3.20.0
python-binance==1.0.19
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.4
pytz==2024.2
regex==2024.9.11
requests==2.32.3
setuptools==75.1.0
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
websockets==13.1
winloop==0.1.6
wsproto==1.2.0
yarl==1.12.1
